# WIP: Improve the speed of searchsorted.