@shoyer 2018-10-10
I think this is missing your changes -- I only see a few lines moved in `binsearch.c.src`. If you edited `binsearch.c` instead, note that that file is autogenerated.

Also, you should remove `site.cfg`.