@charris 2018-10-10
The `site.cfg` file doesn't belong here, you need to be more careful about commits.