@rainwoodman 2018-10-10
Oops I was hoping to get a few comments about the potential effect on the performance; but most were about the PR itselfs. The commits are not cleaned up because this was a WIP -- I was trying to find a way to workaround the header include order bug and get asv working.

It will slightly slow some cases while drastically increase the speed for others. How acceptable is this?