@shoyer 2018-10-10
Feel free to reopen after removing the `site.cfg` file