@shoyer 2018-10-10
> The change was that small!

OK, I see now this works now. Very simple, I like it! I'll ponder more on the performance implications.