@shoyer 2018-10-10
Could you possibly make some lines plots with the most relevant performance comparisons? It's hard to understand a big table of numbers.