@rainwoodman 2018-10-12
Some progress on this:

http://rainwoodman.github.io/numpy/#bench_function_base.Search.time_sorted?branch=fastersearch-omn&branch=fastersearch-baseline&branch=fastersearch-checkfirst&x-axis=needlesize&p-side='l'&p-haystacksize=10000

The O(m+n) algorithm has a significant advantage when the length of the sorted segment in needles array is longer or comparable to a number that is related to the length of the haystack. I think we can find the length of the sorted segment and decide which algorithm to use on the fly.